Allow opening locked directories in some cases
If a workspace/project/library directory is locked by another user or
if we are not sure whether the lock is stale or not, the user can now
override the lock and open the directory anyway. This is especially
useful when running LibrePCB as a Flatpak since due to PID namespacing,
stale locks cannot be detected.